framework_django/treinaweb-django-rest/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.1
django-oauth-toolkit==1.6.3
djangorestframework==3.13.1
idna==3.3
jwcrypto==1.0
mysqlclient==2.1.0
oauthlib==3.1.1
pycparser==2.21
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
wrapt==1.13.3
